+++ src/components/ParameterVariations/VariationBound.tsx
