+++ src/components/Timeline/TimerControls.tsx
