+++ src/components/TimelineBlockStack/ParameterView.tsx
@@ -3,12 +3,12 @@ import { Box, Button, HStack, Text } from "@chakra-ui/react";
 import { useState } from "react";
 import { BsCaretDown, BsCaretUp } from "react-icons/bs";
 import { Block } from "@/src/types/Block";
-import { NewVariationButtons } from "@/src/components/NewVariationButtons";
-import { ParameterVariations } from "@/src/components/ParameterVariations";
+import { NewVariationButtons } from "@/src/components/ParameterVariations/NewVariationButtons";
 import { observer } from "mobx-react-lite";
-import { ParameterValue } from "@/src/components/ParameterValue";
+import { ParameterValue } from "@/src/components/TimelineBlockStack/ParameterValue";
 import { useStore } from "@/src/types/StoreContext";
-import { HeaderRepeat } from "@/src/components/HeaderRepeat";
+import { HeaderRepeat } from "@/src/components/TimelineBlockStack/HeaderRepeat";
+import { ParameterVariations } from "@/src/components/ParameterVariations/ParameterVariations";
 
 type ParameterProps = {
   uniformName: string;
