+++ src/components/TimelineBlockStack/TimelineBlockStack.tsx
@@ -1,6 +1,6 @@
 import { Block } from "@/src/types/Block";
 import { useStore } from "@/src/types/StoreContext";
-import { TimelineBlockBound } from "@/src/components/TimelineBlockBound";
+import { TimelineBlockBound } from "@/src/components/TimelineBlockStack/TimelineBlockBound";
 import { Card } from "@chakra-ui/react";
 import { action, computed } from "mobx";
 import { observer } from "mobx-react-lite";
@@ -14,8 +14,8 @@ import {
 import Draggable from "react-draggable";
 import { DraggableData } from "react-draggable";
 import { DraggableEvent } from "react-draggable";
-import { PatternOrEffectBlock } from "@/src/components/PatternOrEffectBlock";
-import { AddEffectButton } from "@/src/components/AddEffectButton";
+import { PatternOrEffectBlock } from "@/src/components/TimelineBlockStack/PatternOrEffectBlock";
+import { AddEffectButton } from "@/src/components/TimelineBlockStack/AddEffectButton";
 
 type Props = {
   patternBlock: Block;
